Moved all of compare functionality into the Game class. Deleted Compare class and test file.
